requirements.txt
streamlit
langchain
langchain-openai
langchain-community
python-dotenv
langchain-google-genai
faiss-cpu
PyPDF2
sentence-transformers